+++ modules/store/src/main/resources/db/migration/mariadb/V1.6.0__integration_enabled.sql
@@ -1,0 +1,7 @@
+ALTER TABLE `collective`
+ADD COLUMN (`integration_enabled` BOOLEAN);
+
+UPDATE `collective` SET `integration_enabled` = true;
+
+ALTER TABLE `collective`
+MODIFY `integration_enabled` BOOLEAN NOT NULL;